requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
Bottleneck @ file:///C:/ci/bottleneck_1648029530934/work
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.2
gunicorn==20.1.0
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyter-server==1.16.0
jupyterlab==3.3.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib @ file:///C:/ci/matplotlib-suite_1647423638658/work
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
munkres==1.1.4
nbclassic==0.3.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
notebook-shim==0.1.0
numexpr @ file:///C:/ci_310/numexpr_1642084889105/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1649782984482/work
openpyxl==3.0.9
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1650373461903/work
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent==0.18.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
simplejson==3.17.6
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
statsmodels==0.13.2
tenacity==8.0.1
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado @ file:///C:/ci_310/tornado_1642093111997/work
traitlets==5.1.1
tzdata==2022.1
Unipath==1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
whitenoise==6.1.0
widgetsnbextension==3.6.0
wincertstore==0.2
yahoo-finance==1.4.0
